3.x: Makes SecretBundleConfigSource public; improves modification logic (backport of #7421)
